Fixing nil error when working with some windows.
I use zoom (alot) now that we are all working from home. I can't use exwm-mff
with zoom because when I try to launch it some window comes up at nil when
detecting if the mouse is in the window.

I have modified exwm-mff--contains-pointer? to return t if the window is nil.
This prevents both trying to query the windows location, and trying to move it.
Once again allowing me to use exwm-mff with zoom.